hisi_mailbox: allow sending auto-ack messages synchronously
When communicating with LPM3 to request HiFi halt we need to have a
point in time when the command is known to be completed. With
asynchronous command sending there seem to be no way to find such point:
it may be run long after the completion of the request. With synchronous
command sending LPM3 apparently gets confused and does not respond
properly. So there's an artificial mixture of the request
acknowledgement type (auto / manual) and request sending mode
(asynchronous / synchronous). Add the function hisi_rproc_xfer_sync_auto
that sends auto-ack message synchronously.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>